add algorithm
