+++ src/db/keypair/insert.ts
@@ -1,11 +1,14 @@
 import { Keypairs } from "@prisma/client";
 import { createHash } from "crypto";
+import { KeypairAlgorithm } from "../../server/schemas/keypairs";
 import { prisma } from "../client";
 
 export const insertKeypair = async ({
   publicKey,
+  algorithm,
 }: {
   publicKey: string;
+  algorithm: KeypairAlgorithm;
 }): Promise<Keypairs> => {
   const hash = createHash("sha256").update(publicKey).digest("hex");
 
@@ -13,6 +16,7 @@ export const insertKeypair = async ({
     data: {
       hash,
       publicKey,
+      algorithm,
     },
   });
 };
