+++ src/prisma/schema.prisma
@@ -268,6 +268,7 @@ model ChainIndexers {
 model Keypairs {
   hash      String   @id
   publicKey String
+  algorithm String
   createdAt DateTime @default(now())
 
   @@map("keypairs")
