+++ src/utils/cache/keypair.ts
@@ -1,5 +1,5 @@
 import type { Keypairs } from "@prisma/client";
-import { _getKeypair } from "../../db/keypair/get";
+import { getKeypairByPublicKey } from "../../db/keypair/get";
 
 // Cache a public key to the Keypair object, or null if not found.
 export const keypairCache = new Map<string, Keypairs | null>();
@@ -14,7 +14,7 @@ export const getKeypair = async ({
     return cached;
   }
 
-  const keypair = await _getKeypair({ publicKey });
+  const keypair = await getKeypairByPublicKey({ publicKey });
   keypairCache.set(publicKey, keypair);
   return keypair;
 };
